Splits Game.Class
