+++ imj-game-hamazed/src/Imj/Game/Hamazed/World/Draw.hs
@@ -15,6 +15,7 @@ import           Data.Char(intToDigit)
 import qualified Data.Map.Strict as Map(assocs)
 
 import           Imj.Game.Class
+import           Imj.Game.Modify
 import           Imj.Game.Hamazed.World.Types
 import           Imj.Graphics.Render
 import           Imj.Space.Types
