+++ imj-game/imj-game.cabal
@@ -37,6 +37,7 @@ library
                      , Imj.Game.KeysMaps
                      , Imj.Game.Level
                      , Imj.Game.Loop
+                     , Imj.Game.Modify
                      , Imj.Game.Network
                      , Imj.Game.Network.Client
                      , Imj.Game.Network.ClientQueues
