+++ imj-game/src/Imj/Game/State.hs
@@ -20,7 +20,7 @@ import           Imj.Prelude
 import           Prelude(putStr, putStrLn, length)
 
 import           Control.Monad.State.Class(MonadState)
-import           Control.Monad.State.Strict(get, modify')
+import           Control.Monad.State.Strict(get, gets, modify')
 import           Control.Monad.IO.Class(MonadIO)
 import           Control.Monad.Reader.Class(asks)
 import           Data.Text(pack)
